refactor(tooltip): extract CSS time parsing to utils
Move delay parsing into a shared parseCSSTime helper for reuse across components.
