+++ packages/react/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./calendar";
+export * from "./css";
 export * from "./compose";
 export * from "./logger";
 export * from "./dom";
